requirements.txt
aws-cdk.core>=1.150.0
aws-cdk.assertions>=1.150.0
aws-cdk.aws-ec2
aws-cdk.aws-codebuild
aws-cdk.aws-codepipeline
aws-cdk.aws-codedeploy
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-events
aws-cdk.aws-iam
aws-cdk.aws-codecommit
aws-cdk.aws-events-targets
aws-cdk.aws-ecs-patterns
aws-cdk.aws-ecs
aws-cdk.aws-ecr
aws-cdk.aws-secretsmanager